D3D12 fixes
[CL 2857644 by Peter Knepley in UT-Clean branch]

+++ Engine/Source/Runtime/Windows/D3D12RHI/Private/D3D12ConstantBuffer.cpp
@@ -1,4 +1,4 @@
-// Copyright 1998-2014 Epic Games, Inc. All Rights Reserved.
+// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
 
 /*=============================================================================
 D3D12ConstantBuffer.cpp: D3D Constant buffer RHI implementation.
@@ -18,7 +18,7 @@ namespace D3D12RHI
 	const uint32 GConstantBufferSizes[MAX_CONSTANT_BUFFER_SLOTS] =
 	{
 		// CBs must be a multiple of 16
-		Align(MAX_GLOBAL_CONSTANT_BUFFER_SIZE, 16),
+		(uint32)Align(MAX_GLOBAL_CONSTANT_BUFFER_SIZE, 16),
 	};
 }
 
@@ -28,7 +28,7 @@ FD3D12ConstantBuffer::FD3D12ConstantBuffer(FD3D12Device* InParent, uint32 InSize
 	ShadowData(NULL),
 	CurrentUpdateSize(0),
 	TotalUpdateSize(0),
-    FD3D12DeviceChild(InParent)
+	FD3D12DeviceChild(InParent)
 {
 	InitResource();
 }
@@ -46,12 +46,12 @@ void FD3D12ConstantBuffer::InitDynamicRHI()
 	// New circular buffer system for faster constant uploads.  Avoids CopyResource and speeds things up considerably
 	// aligned for best performance
 	ShadowData = (uint8*)FMemory::Malloc(MaxSize, 16);
-	FMemory::Memzero(ShadowData,MaxSize);
+	FMemory::Memzero(ShadowData, MaxSize);
 }
 
 void FD3D12ConstantBuffer::ReleaseDynamicRHI()
 {
-	if(ShadowData)
+	if (ShadowData)
 	{
 		FMemory::Free(ShadowData);
 	}
